Big Note Form refactor
