+++ src/utils.js
@@ -11,3 +11,15 @@ export const titleCase = string => {
   // Join the words back together into a string
   return words.join(" ")
 }
+
+
+export const tagArrayToString = tagArray => {
+  return tagArray.join(', ')
+}
+
+
+export const tagStringToArray = tagString => {
+  const rawNames = tagString.split(',')  // Split on comma into array of raw tag names
+  const trimmedNames = rawNames.map(rawName => rawName.trim())  // Trim leading and trailing whitespace from each tag name
+  return trimmedNames.map(trimmedName => trimmedName.replace(/\s\s+/g, ' '))  // Replace internal whitespace with single space
+}
